refactor(v1): tunnel package owns host_endpoint; server binds via its tunnel
- Relocate `host_endpoint` + its `open_tunnel` retry helper out of
  `runtimes/base.py` into `interception/tunnel/prime.py` (the prime_tunnel
  primitive, alongside `PrimeTunnel`). `PrimeTunnel.expose` uses it directly;
  the tool-serving `reachable_url` (still in runtimes) imports it lazily to
  avoid an import cycle. The host-wide `TUNNEL_LIMITER` stays in
  `runtimes.limiters` (a shared rate-limit resource, used by sandbox creation
  too) and is imported lazily inside `host_endpoint`. So the tunnel package no
  longer imports a runtime *method*.
- `InterceptionServer(tunnel)` binds where its tunnel says (`bind_host` /
  `bind_port`), instead of the pool threading those through as separate kwargs;
  no tunnel = loopback ephemeral (the bare-server default).
